requirements.txt
pyrogram
tgcrypto
PyYAML
